Semantics-New: updates related to points-to analysis
git-svn-id: https://scm.cri.ensmp.fr/svn/validation/trunk@5922 67b5017d-e62d-0410-a871-f72b1c62810f

+++ Semantics-New/pointer02b.c
@@ -1,6 +1,11 @@
-// Check the use and translation of anywhere effect. Same as
-// pointer01.c, but the anywhere effect should be typed and not impact x,
-// y and z.
+/* Check the use and translation of anywhere effect. Same as
+ * pointer01.c, but the anywhere effect should be typed and not impact x,
+ * y and z.
+ *
+ * Bug: the translation of intraprocedural points-to does not occur
+ * and a reference to a points-to stub of "foo", "foo:_p_1" is
+ * preserved in the effects of "main".
+ */
 
 #include <assert.h>
 
@@ -17,7 +22,9 @@ void foo(double *p)
 int main()
 {
   x= 1, y=2, z=3;
-  double *p = (int *)malloc(sizeof(double));
+  // FI: was the meaningless cast to (int *) put here with some purpose
+  // double *p = (int *) malloc(sizeof(double));
+  double *p = (double *) malloc(sizeof(double));
   foo(p);
   return x+y+z; // information about x, y and z should be preserved
 }
